+++ packages/amplify-graphql-schema-generator/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-generator@0.6.2...@aws-amplify/graphql-schema-generator@0.6.3) (2024-01-22)
+
+**Note:** Version bump only for package @aws-amplify/graphql-schema-generator
+
 ## [0.6.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-generator@0.6.1...@aws-amplify/graphql-schema-generator@0.6.2) (2023-12-18)
 
 ### Bug Fixes