+++ packages/amplify-graphql-transformer-core/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-core@2.4.2...@aws-amplify/graphql-transformer-core@2.4.3) (2024-01-22)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-core
+
 ## [2.4.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-core@2.4.1...@aws-amplify/graphql-transformer-core@2.4.2) (2023-12-18)
 
 **Note:** Version bump only for package @aws-amplify/graphql-transformer-core